QuotesAppAPI/quotes-web/requirements.txt
fastapi~=0.89.1
uvicorn
pydantic~=1.10.4
pymongo~=4.3.3
Pillow~=9.4.0
textwrap3
boto3~=1.26.72
botocore~=1.29.72